% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu2_opendf.R
\name{docu2_opendf}
\alias{docu2_opendf}
\title{Get documentation from R data frame.}
\usage{
docu2_opendf(input, languages = "current", style = "both")
}
\arguments{
\item{input}{R data frame (df) or variable from an R data frame (df$var).}

\item{languages}{Select the language in which the descriptions and labels of the data will be displayed
\itemize{
\item By default the language that is set to current is displayed
(\code{languages = "current"}).
\item The default-option chooses either the default language(if labels and descriptions without a language tag exist)
\item Otherwise the current language is displayed.
(\code{languages = "default"}).
\item You can choose to view all available language variants by selecting
(\code{languages = "all"}),
\item or you can select the language by language code, e.g.
\code{languages = "en"}.
}}

\item{style}{Selects where the output should be displayed (console ore viewer). Default is "both"
\itemize{
\item You can choose to display the code only in the console
(\code{style = "console"}).
(\code{style = "print"}).
\item You can choose to display the code only in the viewer
(\code{style = "viewer"}).
(\code{style = "html"}).
}}
}
\value{
Documentation.
}
\description{
Get access to information about the dataset
and variables via the R-Studio Viewer or the web browser.
}
\examples{
# get example data from the opendataformat package
df <- get(data("data_opendf"))

# view documentation about the dataset in the language that is currently set
\dontrun{
docu2_opendf(df)
}

# view information from a selected variable in language "en"
\dontrun{
docu2_opendf(df$bap87, languages = "en")
}

# view dataset information for all available languages
\dontrun{
docu2_opendf(df, languages = "all")
}


}
