% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docu_opendf_old.R
\name{docu_opendf_old}
\alias{docu_opendf_old}
\title{Get documentation from R data frame.}
\usage{
docu_opendf_old(input, languages = "current", variables = "no", style = "html")
}
\arguments{
\item{input}{R data frame (df) or variable from an R data frame (df$var).}

\item{languages}{Select the language in which the descriptions and labels of the data will be displayed
\itemize{
\item By default the language that is set to current is displayed
(\code{languages = "current"}).
\item You can choose to view all available language variants by selecting
(\code{languages = "all"}),
\item or you can select the language by language code, e.g.
\code{languages = "en"}.
}}

\item{variables}{You can choose whether or not to display all the available metadata
(labels and descriptions of the data set and the variables).
\itemize{
\item By default, information describing the dataset is displayed (\code{variables = "no"}).
\item Selecting \code{variables = "yes"} will also display information about
the variables included in the dataset.
}}

\item{style}{You can choose between three display formats:
\itemize{
\item By default the documentation is displayed
as an HTML page (\code{style = "html"}).
\item To view the documentation within the R console, select \code{style = "print"}
\item If you want to take a look at both styles, use \code{style = "all"}.
}}
}
\value{
Documentation.
}
\description{
Get access to information about the dataset
and variables via the R-Studio Viewer or the web browser.
}
\examples{
# get example data from the opendataformat package
df <- get(data("data_opendf"))

\dontrun{
# view documentation about the dataset
docu_opendf_old(df)

# view information from a selected variable
docu_opendf_old(df$bap87)

# view dataset and variable information
docu_opendf_old(df, variables = "yes")

# view information for all available languages
docu_opendf_old(df, languages = "all")

# print information to the R console
docu_opendf_old(df$bap87, style = "print")
}


}
