% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata_opendf.R
\name{getmetadata_opendf}
\alias{getmetadata_opendf}
\title{Get variable labels or other metadata from a data frame in opendataformat.}
\usage{
getmetadata_opendf(input, type, language = "active")
}
\arguments{
\item{input}{R data frame (df) or variable from an R data frame (df$var).}

\item{language}{Select the language in which the labels of the variables
will be displayed. If no language is selected, the current/active language
of the data frame will be used.
\itemize{
\item By default the language that is set to current is displayed (\code{language = "current"}).
\item You can select the language by language code, e.g. \code{language = "en"}.
}}

\item{valuelabels}{You can choose to display the value labels instead of
the variable label, if the input is a variable by setting \code{valuelabels=TRUE}
or \code{valuelabels="yes"}. If the input is a dataset, this argument is
ignored.}

\item{retrieve}{You can choose to display another attribute/metadata instead
of the variable label(s). Possible options are "description", "url", "type"
or "languages".If \code{valuelabels=TRUE} or \code{valuelabels="yes"} and
the input is a variable, this argument is ignored.}
}
\value{
Documentation.
}
\description{
Get access to information about the dataset
and variables via the R-Studio Viewer or the web browser.
}
\examples{
# get example data from the opendataformat package
df <- get(data("data_opendf"))
# view the variable labels for all variables in English
getmetadata_opendf(input = df, type="label", language = "en")

# view the value labels for variable bap87 in English
getmetadata_opendf(input = df$bap87, type = "valuelabel", language = "en")

# view the description for variable bap87 in English
getmetadata_opendf(input = df$bap87, type = "description", language = "en")

}
