% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendataformat-package.R
\docType{package}
\name{opendataformat-package}
\alias{opendataformat-package}
\alias{opendataformat}
\title{Open Data Format}
\description{
The package is designed to support the use of the open data format.
For this purpose, three main functions have been developed:
\subsection{read_opendf()}{

Import data from the Open Data Format to an R data frame.
}

\subsection{write_opendf()}{

Export data from an R data frame to the open data format.
}

\subsection{docu_opendf()}{

Get access to information about the dataset
and variables via the R-Studio Viewer or the web browser.
}
}
\seealso{
More information about the Open Data Format specification and
data examples are available here:
\url{https://git.soep.de/opendata/}
}
\author{
Tom Hartl (\email{thartl@diw.de}),
Claudia Saalbach (\email{csaalbach@diw.de}),
Xiaoyao Han (\email{xhan@diw.de}),
Knut Wenzig (\email{kwenzig@diw.de})

Other Contributors: KonsortSWD/NFDI, DIW Berlin
}
