% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2csv.R
\name{r2csv}
\alias{r2csv}
\title{RData to CSV}
\usage{
r2csv(input, output, languages, variables, export_data)
}
\arguments{
\item{input}{R dataframe object.}

\item{output}{path to directory for saving the CSV files.}

\item{languages}{select language option: "all" = all available languages, "default" = language that is set
to default, or language code for exmaple "de" or "en".}

\item{variables}{Choose, if all dataset variables should be converted.}

\item{export_data}{choose whether the dataset CSV should be copied to the output directory ("yes") or not ("no").}
}
\value{
CSV files holding the data and the metadata.
}
\description{
The r2csv function cssonverts a R data frame to the CSV open data format.
}
