% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opendf.R
\name{read_opendf}
\alias{read_opendf}
\title{Read data specified as Open Data Format.}
\usage{
read_opendf(file, languages = "all")
}
\arguments{
\item{file}{Path to ZIP file containing data and metadata specified
as Open Data Format.}

\item{languages}{Select the language in which the descriptions and labels of the data will be imported.
\itemize{
\item By default all available language variants are imported (\code{languages = "all"}).
\item You can also choose to import only the default language (\code{languages = "default"}),
\item or you can select the language by language code, e.g. \code{languages = "en"}.
}}
}
\value{
R dataframe with attributes including dataset and variable information.
}
\description{
Import data from the Open Data Format to an R data frame.
}
\examples{
# get path to example data from the opendataformat package (data.zip)
path <- system.file("extdata", "data.zip", package="opendataformat")
path

# read example data specified as Open Data Format from ZIP file
df <- read_opendf(input = path)
df
attributes(df)
attributes(df$bap87)

# read example data with language selection
df <- read_opendf(input = path, languages = "de")
attributes(df$bap87)

}
