% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_opendf.R
\name{write_opendf}
\alias{write_opendf}
\title{Write R data frame to the Open Data Format.}
\usage{
write_opendf(
  input,
  output,
  languages = "all",
  variables = "yes",
  export_data = "yes"
)
}
\arguments{
\item{input}{R data frame (df).}

\item{output}{Path to ZIP file.}

\item{languages}{Select the language in which the descriptions and labels of the data will be exported
\itemize{
\item By default all available language variants are exported
(\code{languages = "all"}).
\item You can also choose to export only the default language
(\code{languages = "default"}),
\item or you can select the language by language code, e.g.
\code{languages = "en"}.
}}

\item{variables}{If you are exporting the entire dataset, you can choose whether or not to
export all available metadata (labels and descriptions of the dataset and the variables).
\itemize{
\item By default, all metadata is exported: information describing the dataset
itself, as well as information describing all variables in
the dataset (\code{variables = "yes"}).
\item If you set \code{variables = "no"}, only the information describing the
dataset is exported.
}}

\item{export_data}{Choose, if you want to export the file that holds the
data (data.csv).
\itemize{
\item By default the data and metadata are exported (\code{export_data = "yes"}).
\item To export only metadata and no data, select \code{export_data = "no"}
}}
}
\value{
ZIP file and unzipped directory containing the data as CSV file and
the metadata as XML file (DDI Codebook 2.5.).
}
\description{
Export data from an R data frame to a ZIP file that stores
the data as Open Data Format.
}
\examples{
# get example data from the opendataformat package
df <- get(data("data_opendf"))
df

# write R data frame with attributes to the file my_data.zip specified
# as Open Data Format.
\dontrun{
write_opendf(input = df, output = "my_data")
}

# write variable from R data frame with attributes to the file my_data.zip specified
# as Open Data Format.
\dontrun{
write_opendf(input = df$bap87, output = "my_data")
}

# write R data frame with attributes to the file my_data.zip
# with selected language.
\dontrun{
write_opendf(input = df, output = "my_data", languages = "en")
}

# write R data frame with attributes to the file my_data.zip but only
# metadata, no data.
\dontrun{
write_opendf(input = df, output = "my_data", export_data = "no")
}


}
