% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opendf2.R
\name{read_opendf2}
\alias{read_opendf2}
\title{Read data specified as Open Data Format.}
\usage{
read_opendf2(file, languages = "all", nrows = Inf, skip = 0, variables = NULL)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
By default all available language variants are imported (\code{languages = "all"}).}

\item{languages}{integer: the maximum number of rows to read in. Negative and other invalid values are ignored.}

\item{nrows}{Maximum number of lines to read.}

\item{skip}{Select the number of rows to be skipped (without the column names).}

\item{variables}{Columns to include in the results. You can use the same mini-language as dplyr::select() to refer to the columns by name. Use c() to use more than one selection expression.
Although this usage is less common, col_select also accepts a numeric column index. See ?tidyselect::language for full details on the selection language.}
}
\value{
R dataframe with attributes including dataset and variable information.
}
\description{
Import data from the Open Data Format to an R data frame.
}
\examples{
# get path to example data from the opendataformat package (data.zip)
path <- system.file("extdata", "data.zip", package="opendataformat")
path

# read example data specified as Open Data Format from ZIP file
df <- read_opendf2(file = path)
df
attributes(df)
attributes(df$bap87)

# read example data with language selection
df <- read_opendf(file = path, languages = "de")
attributes(df$bap87)

}
